+++ src/reducers/index.js
@@ -1,7 +1,11 @@
 import { combineReducers } from 'redux';
 
-import user from './user';
+import auth from './auth';
+import profile from './profile';
+import pokedex from './pokedex';
 
 export default combineReducers({
-  user,
+  auth,
+  profile,
+  pokedex,
 });